item/actions/effects: merge ITEM_ACTION_FLOOR_SHAKE handler
