+++ src/tr1/game/item_actions.c
@@ -1,6 +1,5 @@
 #include "game/item_actions/finish_level.h"
 #include "game/item_actions/flipmap.h"
-#include "game/item_actions/floor_shake.h"
 #include "game/item_actions/powerup.h"
 #include "game/item_actions/raising_block.h"
 #include "game/item_actions/sand.h"
@@ -16,7 +15,6 @@ void Item_ActionRunLegacy(ITEM_TRX_ACTION action_id, ITEM *item)
 {
     static M_FUNC m_Actions[] = {
         [ITEM_ACTION_TURN_180] = ItemAction_Turn180,
-        [ITEM_ACTION_FLOOR_SHAKE] = ItemAction_FloorShake,
         [ITEM_ACTION_FINISH_LEVEL] = ItemAction_FinishLevel,
         [ITEM_ACTION_RAISING_BLOCK] = ItemAction_RaisingBlock,
         [ITEM_ACTION_STAIRS_TO_SLOPE] = ItemAction_Stairs2Slope,
