+++ src/tr1/meson.build
@@ -115,7 +115,6 @@ sources = [
   'game/item_actions.c',
   'game/item_actions/finish_level.c',
   'game/item_actions/flipmap.c',
-  'game/item_actions/floor_shake.c',
   'game/item_actions/powerup.c',
   'game/item_actions/raising_block.c',
   'game/item_actions/sand.c',
